refactor(handlers): share read-only planning pipeline
M3-REF-004a (#524)
